rename shared-service to section-service
